refactor(ui): finish SettingsDrawer split (loaders + savers hooks)
Take SettingsDrawer.tsx from 1149 to 711 lines, clearing the 800-line
red-flag budget. Two new hooks own the per-section fetch and save
callbacks that previously lived inline in the drawer.

- hooks/useSettingsDrawerLoaders.ts (~400): every per-section fetch
  callback (thresholds, IP, tests, iperf suggestions, wifi, network
  discovery, snmp, link, cable, logs) plus the open-time
  initial-load orchestration useEffect that previously fanned out
  via ten dependencies in the drawer
- hooks/useSettingsDrawerSavers.ts (~225): the seven section save
  callbacks (thresholds, tests, wifi, link, cable, network discovery,
  snmp) - each pure-fetch with no UI rendering

The drawer keeps the state objects (still numerous) but no longer
owns the per-call networking. saveIpSettings stays inline because
it's still manual-apply via the user-facing button. Pure relocation,
no behavior change.
